배송 매니저 생성 로직
