+++ delivery-service/src/main/java/org/sixpang/deliveryservice/domain/repository/DeliveryManagerRepository.java
@@ -1,0 +1,11 @@
+package org.sixpang.deliveryservice.domain.repository;
+
+import org.sixpang.deliveryservice.domain.model.entity.HubDeliveryManager;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface DeliveryManagerRepository extends JpaRepository<HubDeliveryManager, UUID> {
+    Optional<HubDeliveryManager> findFirstByHubIdAndStatusOrderByDeliverySequenceAsc(UUID deliveryManagerId);
+
+}
